Fix 108 findings from second fresh exhaustive codebase review
Addresses every confirmed finding from a fresh multi-agent review across
all subsystems and 5 backends (2 critical, 30 high, 37 medium, 39 low; the
1 adversarially-rejected finding excluded). Full correct fixes only — no
workarounds, no CPU fallbacks for GPU backends.

Critical (both security):
- scatter/scatter_add/scatter_reduce (+gather/put) OOB on index extent
  exceeding input extent on a non-scatter dim.
- Lite runtime weight view: validate numel*dtype_size against the
  bounds-checked weight_nbytes before from_blob (OOB on untrusted .tzlite).

Highlights — several were regressions from the prior fix wave:
- CUDA softmax/log_softmax backward half/bf16 accumulators (mirror forward
  AccumType); SDPA per-device workspace+stream; cuSOLVER half-narrowing;
  64-bit warp-match pooling; ConvTranspose3d/cdist/histogram double accum.
- ROCm fused sgd/adam/adam_atan2 grid-stride tail; Conv3d bias-grad double
  accumulator; roi_align empty-map guard; cross-device peer copy.
- OneAPI max/min/argmax strides; rfft/irfft length; scatter index-stride OOB.
- Vulkan argmax dim=-1; dispatchFill Int64; F16/BF16 atomic RMW; and the
  sparse_trsm cross-workgroup deadlock REGRESSION rewritten single-workgroup.
- Autograd FusedLinearReLU grad transpose, SVD/Eigh degenerate clamp, jvp
  multi-output, gradcheck half-widen; optimizer null-param buffer alignment;
  INT4 QConfig guard; JIT multi-axis reduction; lazy reshape validation;
  async transfer contiguify; sparse CSC/BSR mul; Future take() copy;
  torch_pickle ZIP64 overflow-safe bounds; distributed socket hardening.

Two prior workarounds upgraded to real fixes:
- Fused CUDA optimizer keeps double-precision hyperparameters for Float64.
- ONNX GRU linear_before_reset=0 (ONNX default) properly implemented in
  GRUCell with a mode flag (default = unchanged mode-1) instead of throwing.

Whole library builds: main (core + CPU + Python + all tests/examples) and
all 5 backend builds link clean; affected test suites pass.

+++ src/backends/cuda/kernels/conv3d.cu
@@ -965,7 +965,11 @@ auto conv_transpose3d_forward_kernel(
 
     switch (input.dtype()) {
         case DType::Float32:
-            launch_conv_transpose3d_forward<float, float>(
+            // Accumulate Float32 ConvTranspose3d in double, matching the Conv3d
+            // decision (a float accumulator drifts O(sqrt(N)) ULPs over the
+            // Cin_per_g*kD*kH*kW reduction and pushes the cross-backend rtol=1e-3
+            // parity test over threshold).
+            launch_conv_transpose3d_forward<float, double>(
                 input, weight, bias, output, N, Cin, D, H, W, Cout, Cout_per_g,
                 kD, kH, kW, oD, oH, oW, s, p, d, groups, stream);
             break;
@@ -1023,7 +1027,8 @@ auto conv_transpose3d_backward_input_kernel(
 
     switch (input.dtype()) {
         case DType::Float32:
-            launch_conv_transpose3d_backward_input<float, float>(
+            // Double accumulator for Float32 (cross-backend parity); see fwd note.
+            launch_conv_transpose3d_backward_input<float, double>(
                 grad_output, weight, grad_input, N, Cin, D, H, W, Cout, Cout_per_g,
                 kD, kH, kW, oD, oH, oW, s, p, d, groups, stream);
             break;
@@ -1077,7 +1082,8 @@ auto conv_transpose3d_backward_weight_kernel(
 
     switch (input.dtype()) {
         case DType::Float32:
-            launch_conv_transpose3d_backward_weight<float, float>(
+            // Double accumulator for Float32 (cross-backend parity); see fwd note.
+            launch_conv_transpose3d_backward_weight<float, double>(
                 grad_output, input, grad_weight, N, Cin, D, H, W, Cout, Cout_per_g,
                 kD, kH, kW, oD, oH, oW, s, p, d, groups, stream);
             break;
